+++ src/vue2/index.js
@@ -2,3 +2,5 @@ export {
   makeMultiValueRenderer,
   default as MultiValueRenderers
 } from './MultiValueTableRenderer'
+
+export { getAggregatorNumInputs } from '../core'